LayoutTests/storage/indexeddb/modern/keypath-basic-expected.txt

ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: object put SUCCESS - foo1
ALERT: array put SUCCESS - foo2
ALERT: Second object put SUCCESS - baz1
ALERT: Second array put SUCCESS - baz2
ALERT: Initial upgrade versionchange transaction complete
ALERT: Done
This test creates some object stores with keypaths.
It then puts some values in them.
It makes sure the keys used are as expected.
